feat(core-agent): extract pure runtime logic from stage-ui into packages/core-agent
