+++ packages/core-agent/src/contracts/context-port.ts
@@ -1,0 +1,7 @@
+import type { ContextMessage } from '../types/chat'
+
+export interface AgentContextPort {
+  ingest: (envelope: ContextMessage) => void
+  snapshot: () => Record<string, ContextMessage[]>
+  reset: () => void
+}
