+++ packages/core-agent/src/contracts/llm-port.ts
@@ -1,0 +1,8 @@
+import type { ChatProvider } from '@xsai-ext/providers/utils'
+import type { Message } from '@xsai/shared-chat'
+
+import type { StreamOptions } from '../types/llm'
+
+export interface AgentLLMPort {
+  stream: (model: string, chatProvider: ChatProvider, messages: Message[], options?: StreamOptions) => Promise<void>
+}
