+++ packages/core-agent/src/contracts/session-port.ts
@@ -1,0 +1,8 @@
+import type { ChatHistoryItem } from '../types/chat'
+
+export interface AgentSessionPort {
+  ensureSession: (sessionId: string) => void
+  getSessionMessages: (sessionId: string) => ChatHistoryItem[]
+  appendSessionMessage: (sessionId: string, message: ChatHistoryItem) => void
+  getSessionGeneration: (sessionId: string) => number
+}
