+++ packages/core-agent/src/contracts/stream-port.ts
@@ -1,0 +1,6 @@
+import type { StreamingAssistantMessage } from '../types/chat'
+
+export interface AgentForegroundStreamPort {
+  patch: (message: StreamingAssistantMessage) => void
+  reset: () => void
+}
