+++ packages/core-agent/src/index.ts
@@ -1,0 +1,40 @@
+export type { AgentContextPort } from './contracts/context-port'
+
+export type { ChatHookRegistry } from './contracts/hook-types'
+
+export type { AgentLLMPort } from './contracts/llm-port'
+export type { AgentSessionPort } from './contracts/session-port'
+export type { AgentForegroundStreamPort } from './contracts/stream-port'
+export { createChatHooks } from './runtime/agent-hooks'
+export type { ContextHistoryEntry, ContextRegistry } from './runtime/context-registry'
+
+export { createContextRegistry } from './runtime/context-registry'
+export {
+  isToolRelatedError,
+  modelKey,
+  sanitizeMessages,
+  streamFrom,
+  streamOptionsToolsCompatibilityOk,
+} from './runtime/llm-service'
+export { mergeLoadedSessionMessages } from './session/merge-loaded-session-messages'
+export type {
+  ChatAssistantMessage,
+  ChatHistoryItem,
+  ChatMessage,
+  ChatSlices,
+  ChatSlicesText,
+  ChatSlicesToolCall,
+  ChatSlicesToolCallResult,
+  ChatStreamEvent,
+  ChatStreamEventContext,
+  ContextMessage,
+  ErrorMessage,
+  StreamingAssistantMessage,
+} from './types/chat'
+
+export type {
+  BuiltinToolsResolver,
+  StreamEvent,
+  StreamFromOptions,
+  StreamOptions,
+} from './types/llm'
