+++ packages/core-agent/src/types/llm.ts
@@ -1,0 +1,30 @@
+import type { ChatProvider } from '@xsai-ext/providers/utils'
+import type { CommonContentPart, CompletionToolCall, CompletionToolResult, Message, Tool } from '@xsai/shared-chat'
+
+export type StreamEvent
+  = | { type: 'text-delta', text: string }
+    | ({ type: 'finish' } & any)
+    | ({ type: 'tool-call' } & CompletionToolCall)
+    | (CompletionToolResult & { type: 'tool-error' })
+    | { type: 'tool-result', toolCallId: string, result?: string | CommonContentPart[] }
+    | { type: 'error', error: any }
+
+export interface StreamOptions {
+  abortSignal?: AbortSignal
+  headers?: Record<string, string>
+  onStreamEvent?: (event: StreamEvent) => void | Promise<void>
+  toolsCompatibility?: Map<string, boolean>
+  supportsTools?: boolean
+  waitForTools?: boolean
+  tools?: Tool[] | (() => Promise<Tool[] | undefined>)
+}
+
+export type BuiltinToolsResolver = (model: string, chatProvider: ChatProvider) => Promise<Tool[]>
+
+export interface StreamFromOptions {
+  model: string
+  chatProvider: ChatProvider
+  messages: Message[]
+  options?: StreamOptions
+  builtinToolsResolver?: BuiltinToolsResolver
+}
