+++ packages/core-agent/tsdown.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'tsdown'
+
+export default defineConfig({
+  entry: [
+    'src/index.ts',
+  ],
+  dts: true,
+})
