+++ packages/stage-ui/src/stores/exports.contract.test.ts
@@ -31,6 +31,8 @@ describe('stage-ui exports contract', () => {
       './constants/*',
       './libs',
       './libs/*',
+      './libs/inference',
+      './libs/inference/adapters/*',
       './stores',
       './stores/*',
       './stores/analytics',
